RichMenuContentsコンポーネントの作成
