+++ src/components/organisms/RichMenu/RichMenuContents/RichMenuContents.stories.tsx
@@ -1,0 +1,17 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import { RichMenuContents } from "./RichMenuContents";
+
+export default {
+  component: RichMenuContents,
+  parameters: {
+    controls: { expanded: true },
+  },
+  argTypes: {},
+} satisfies Meta<typeof RichMenuContents>;
+
+export const Default: StoryObj<typeof RichMenuContents> = {
+  args: {},
+};
+
+
