+++ src/components/organisms/RichMenu/RichMenuContents/index.ts
@@ -1,0 +1,1 @@
+export * from './RichMenuContents';
