+++ src/components/organisms/RichMenu/RichMenuContents/type.ts
@@ -1,0 +1,5 @@
+import { RichMenuTypeKey } from "../type";
+
+export type RichMenuContentsProps = {
+  richMenuType: RichMenuTypeKey;
+};
